AMP-10498
git-svn-id: file:///svn/branches/AMP_1_16_RELEASE@15823 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/digijava/module/dataExchange/engine/DEImportBuilder.java
@@ -633,7 +633,7 @@ public class DEImportBuilder {
 				
 				AmpActivityProgram ampActivityProgram = new AmpActivityProgram();
 				ampActivityProgram.setActivity(activity);
-				ampActivityProgram.setProgramPercentage((new Float(idmlProgram.getPercentage())));
+				ampActivityProgram.setProgramPercentage((new Float(idmlProgram.getPercentage()).longValue()));
 				ampActivityProgram.setProgram(ampTheme);
 				
 				ArrayList<AmpActivityProgramSettings> allClassifConfigs = (ArrayList<AmpActivityProgramSettings>) getAllAmpActivityProgramSettings();
